Rename FormButton to Button
